# diagrams/CoreService_swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "UniGate.Core API",
    "version": "v1"
  },
  "paths": {
...
    "/api/programs/{id}/apply": {
      "post": {
        "tags": [
          "Programs"
        ],
        "summary": "Apply for an education program",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/me/chosen-programs/{id}": {
      "put": {
        "tags": [
          "Programs"
        ],
        "summary": "Change the priority of an education program",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeProgramPriorityDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeProgramPriorityDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeProgramPriorityDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeProgramPriorityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Programs"
        ],
        "summary": "Remove education program from chosen",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/applicants/{applicantId}/chosen-programs/{programId}": {
      "delete": {
        "tags": [
          "Programs"
        ],
        "summary": "Remove education of applicant program from chosen",
        "parameters": [
          {
            "name": "applicantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "programId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ChangeProgramPriorityDto": {
        "type": "object",
        "properties": {
          "priority": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}